CAM: cleanup: add test for DetachedDocumentObject and deduplicate code that interacts with freecad documents
